% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.r
\name{AmberClient}
\alias{AmberClient}
\title{Default operations}
\description{
swagger.Default
}
\section{Methods}{

\describe{

delete_sensor Delete a sensor instance


get_amber_summary Get the JSON block of the amber image


get_config Get the current configuration of a sensor instance


get_pretrain_state Get status of pretrain operation


get_root_cause Get root cause analysis information from a sensor


get_sensor Get basic information about a sensor instance


list_sensors List all sensors for this user


get_status Get analytic information from a sensor


get_version Retrieves API version information


configure_sensor Apply configuration to a sensor instance


post_oauth2 Request a bearer token using Amber account credentials


pretrain_sensor Pretrain a sensor using historical data


create_sensor Create a new a sensor instance


stream_sensor Stream data to a sensor


configure_fusion Update configuration for a sensor instance for fusion data


enable_learning set new graduation requirements and turns on learning


update_label Update label for a sensor instance


stream_fusion Stream data to a sensor fusion vector

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AmberClient$new()}}
\item \href{#method-list_sensors}{\code{AmberClient$list_sensors()}}
\item \href{#method-get_sensor}{\code{AmberClient$get_sensor()}}
\item \href{#method-create_sensor}{\code{AmberClient$create_sensor()}}
\item \href{#method-update_label}{\code{AmberClient$update_label()}}
\item \href{#method-configure_sensor}{\code{AmberClient$configure_sensor()}}
\item \href{#method-configure_fusion}{\code{AmberClient$configure_fusion()}}
\item \href{#method-enable_learning}{\code{AmberClient$enable_learning()}}
\item \href{#method-get_config}{\code{AmberClient$get_config()}}
\item \href{#method-delete_sensor}{\code{AmberClient$delete_sensor()}}
\item \href{#method-stream_sensor}{\code{AmberClient$stream_sensor()}}
\item \href{#method-stream_fusion}{\code{AmberClient$stream_fusion()}}
\item \href{#method-get_status}{\code{AmberClient$get_status()}}
\item \href{#method-pretrain_sensor}{\code{AmberClient$pretrain_sensor()}}
\item \href{#method-get_pretrain_state}{\code{AmberClient$get_pretrain_state()}}
\item \href{#method-get_root_cause}{\code{AmberClient$get_root_cause()}}
\item \href{#method-get_version}{\code{AmberClient$get_version()}}
\item \href{#method-clone}{\code{AmberClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$new(
  license_id = "default",
  license_file = "~/.Amber.license",
  verify = TRUE,
  cert = NULL,
  timeout = 360
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_sensors"></a>}}
\if{latex}{\out{\hypertarget{method-list_sensors}{}}}
\subsection{Method \code{list_sensors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$list_sensors(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sensor"></a>}}
\if{latex}{\out{\hypertarget{method-get_sensor}{}}}
\subsection{Method \code{get_sensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$get_sensor(sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_sensor"></a>}}
\if{latex}{\out{\hypertarget{method-create_sensor}{}}}
\subsection{Method \code{create_sensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$create_sensor(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_label"></a>}}
\if{latex}{\out{\hypertarget{method-update_label}{}}}
\subsection{Method \code{update_label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$update_label(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-configure_sensor"></a>}}
\if{latex}{\out{\hypertarget{method-configure_sensor}{}}}
\subsection{Method \code{configure_sensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$configure_sensor(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-configure_fusion"></a>}}
\if{latex}{\out{\hypertarget{method-configure_fusion}{}}}
\subsection{Method \code{configure_fusion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$configure_fusion(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-enable_learning"></a>}}
\if{latex}{\out{\hypertarget{method-enable_learning}{}}}
\subsection{Method \code{enable_learning()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$enable_learning(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_config}{}}}
\subsection{Method \code{get_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$get_config(sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_sensor"></a>}}
\if{latex}{\out{\hypertarget{method-delete_sensor}{}}}
\subsection{Method \code{delete_sensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$delete_sensor(sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stream_sensor"></a>}}
\if{latex}{\out{\hypertarget{method-stream_sensor}{}}}
\subsection{Method \code{stream_sensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$stream_sensor(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stream_fusion"></a>}}
\if{latex}{\out{\hypertarget{method-stream_fusion}{}}}
\subsection{Method \code{stream_fusion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$stream_fusion(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$get_status(sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pretrain_sensor"></a>}}
\if{latex}{\out{\hypertarget{method-pretrain_sensor}{}}}
\subsection{Method \code{pretrain_sensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$pretrain_sensor(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pretrain_state"></a>}}
\if{latex}{\out{\hypertarget{method-get_pretrain_state}{}}}
\subsection{Method \code{get_pretrain_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$get_pretrain_state(sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_root_cause"></a>}}
\if{latex}{\out{\hypertarget{method-get_root_cause}{}}}
\subsection{Method \code{get_root_cause()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$get_root_cause(sensor_id, cluster_id, pattern, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_version"></a>}}
\if{latex}{\out{\hypertarget{method-get_version}{}}}
\subsection{Method \code{get_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$get_version(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmberClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
