% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultApi.r
\name{DefaultApi}
\alias{DefaultApi}
\title{Default operations}
\description{
swagger.Default
}
\section{Methods}{

\describe{

delete_sensor Delete a sensor instance


get_amber_summary Get the JSON block of the amber image


get_config Get the current configuration of a sensor instance


get_pretrain Get status of pretrain operation


get_root_cause Get root cause analysis information from a sensor


get_sensor Get basic information about a sensor instance


get_sensors List all sensors for this user


get_status Get analytic information from a sensor


get_version Retrieves API version information


post_config Apply configuration to a sensor instance


post_oauth2 Request a bearer token using Amber account credentials


post_pretrain Pretrain a sensor using historical data


post_sensor Create a new a sensor instance


post_stream Stream data to a sensor


put_config Update configuration for a sensor instance


put_sensor Update label for a sensor instance


put_stream Stream data to a sensor fusion vector

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DefaultApi$new()}}
\item \href{#method-delete_sensor}{\code{DefaultApi$delete_sensor()}}
\item \href{#method-get_amber_summary}{\code{DefaultApi$get_amber_summary()}}
\item \href{#method-get_config}{\code{DefaultApi$get_config()}}
\item \href{#method-get_pretrain}{\code{DefaultApi$get_pretrain()}}
\item \href{#method-get_root_cause}{\code{DefaultApi$get_root_cause()}}
\item \href{#method-get_sensor}{\code{DefaultApi$get_sensor()}}
\item \href{#method-get_sensors}{\code{DefaultApi$get_sensors()}}
\item \href{#method-get_status}{\code{DefaultApi$get_status()}}
\item \href{#method-get_version}{\code{DefaultApi$get_version()}}
\item \href{#method-post_config}{\code{DefaultApi$post_config()}}
\item \href{#method-post_oauth2}{\code{DefaultApi$post_oauth2()}}
\item \href{#method-post_pretrain}{\code{DefaultApi$post_pretrain()}}
\item \href{#method-post_sensor}{\code{DefaultApi$post_sensor()}}
\item \href{#method-post_stream}{\code{DefaultApi$post_stream()}}
\item \href{#method-put_config}{\code{DefaultApi$put_config()}}
\item \href{#method-put_sensor}{\code{DefaultApi$put_sensor()}}
\item \href{#method-put_stream}{\code{DefaultApi$put_stream()}}
\item \href{#method-clone}{\code{DefaultApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_sensor"></a>}}
\if{latex}{\out{\hypertarget{method-delete_sensor}{}}}
\subsection{Method \code{delete_sensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$delete_sensor(sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_amber_summary"></a>}}
\if{latex}{\out{\hypertarget{method-get_amber_summary}{}}}
\subsection{Method \code{get_amber_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$get_amber_summary(sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_config}{}}}
\subsection{Method \code{get_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$get_config(sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pretrain"></a>}}
\if{latex}{\out{\hypertarget{method-get_pretrain}{}}}
\subsection{Method \code{get_pretrain()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$get_pretrain(sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_root_cause"></a>}}
\if{latex}{\out{\hypertarget{method-get_root_cause}{}}}
\subsection{Method \code{get_root_cause()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$get_root_cause(sensor_id, cluster_id, pattern, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sensor"></a>}}
\if{latex}{\out{\hypertarget{method-get_sensor}{}}}
\subsection{Method \code{get_sensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$get_sensor(sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sensors"></a>}}
\if{latex}{\out{\hypertarget{method-get_sensors}{}}}
\subsection{Method \code{get_sensors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$get_sensors(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-get_status}{}}}
\subsection{Method \code{get_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$get_status(sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_version"></a>}}
\if{latex}{\out{\hypertarget{method-get_version}{}}}
\subsection{Method \code{get_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$get_version(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_config"></a>}}
\if{latex}{\out{\hypertarget{method-post_config}{}}}
\subsection{Method \code{post_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$post_config(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_oauth2"></a>}}
\if{latex}{\out{\hypertarget{method-post_oauth2}{}}}
\subsection{Method \code{post_oauth2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$post_oauth2(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_pretrain"></a>}}
\if{latex}{\out{\hypertarget{method-post_pretrain}{}}}
\subsection{Method \code{post_pretrain()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$post_pretrain(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_sensor"></a>}}
\if{latex}{\out{\hypertarget{method-post_sensor}{}}}
\subsection{Method \code{post_sensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$post_sensor(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_stream"></a>}}
\if{latex}{\out{\hypertarget{method-post_stream}{}}}
\subsection{Method \code{post_stream()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$post_stream(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_config"></a>}}
\if{latex}{\out{\hypertarget{method-put_config}{}}}
\subsection{Method \code{put_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$put_config(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_sensor"></a>}}
\if{latex}{\out{\hypertarget{method-put_sensor}{}}}
\subsection{Method \code{put_sensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$put_sensor(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_stream"></a>}}
\if{latex}{\out{\hypertarget{method-put_stream}{}}}
\subsection{Method \code{put_stream()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$put_stream(body, sensor_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
