% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConfigResponse.r
\name{GetConfigResponse}
\alias{GetConfigResponse}
\title{GetConfigResponse Class}
\description{
GetConfigResponse Class

GetConfigResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{anomalyHistoryWindow}}{}

\item{\code{learningRateNumerator}}{}

\item{\code{learningRateDenominator}}{}

\item{\code{learningMaxClusters}}{}

\item{\code{learningMaxSamples}}{}

\item{\code{featureCount}}{}

\item{\code{streamingWindowSize}}{}

\item{\code{features}}{}

\item{\code{percentVariation}}{}

\item{\code{samplesToBuffer}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GetConfigResponse$new()}}
\item \href{#method-toJSON}{\code{GetConfigResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{GetConfigResponse$fromJSON()}}
\item \href{#method-toJSONString}{\code{GetConfigResponse$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GetConfigResponse$fromJSONString()}}
\item \href{#method-clone}{\code{GetConfigResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetConfigResponse$new(
  anomalyHistoryWindow,
  learningRateNumerator,
  learningRateDenominator,
  learningMaxClusters,
  learningMaxSamples,
  featureCount,
  streamingWindowSize,
  features,
  percentVariation,
  samplesToBuffer
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetConfigResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetConfigResponse$fromJSON(GetConfigResponseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetConfigResponse$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetConfigResponse$fromJSONString(GetConfigResponseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetConfigResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
