% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStatusResponse.r
\name{GetStatusResponse}
\alias{GetStatusResponse}
\title{GetStatusResponse Class}
\description{
GetStatusResponse Class

GetStatusResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pca}}{}

\item{\code{clusterGrowth}}{}

\item{\code{clusterSizes}}{}

\item{\code{anomalyIndexes}}{}

\item{\code{frequencyIndexes}}{}

\item{\code{distanceIndexes}}{}

\item{\code{totalInferences}}{}

\item{\code{numClusters}}{}

\item{\code{anomalyThreshold}}{}

\item{\code{state}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GetStatusResponse$new()}}
\item \href{#method-toJSON}{\code{GetStatusResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{GetStatusResponse$fromJSON()}}
\item \href{#method-toJSONString}{\code{GetStatusResponse$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GetStatusResponse$fromJSONString()}}
\item \href{#method-clone}{\code{GetStatusResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStatusResponse$new(
  pca,
  clusterGrowth,
  clusterSizes,
  anomalyIndexes,
  frequencyIndexes,
  distanceIndexes,
  totalInferences,
  numClusters,
  anomalyThreshold,
  state
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStatusResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStatusResponse$fromJSON(GetStatusResponseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStatusResponse$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStatusResponse$fromJSONString(GetStatusResponseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetStatusResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
