% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSummaryResponse.r
\name{GetSummaryResponse}
\alias{GetSummaryResponse}
\title{GetSummaryResponse Class}
\description{
GetSummaryResponse Class

GetSummaryResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{MagicNumber}}{}

\item{\code{VersionNumber}}{}

\item{\code{m_Nano}}{}

\item{\code{m_BufferStats}}{}

\item{\code{m_Autotune}}{}

\item{\code{m_AutotuningElbowClusterCounts}}{}

\item{\code{m_AutotuningElbowPVArray}}{}

\item{\code{m_StreamingParameters}}{}

\item{\code{m_AmberStatus}}{}

\item{\code{m_Training}}{}

\item{\code{m_AnomalyThreshold}}{}

\item{\code{m_AmberWarningCriticalValue}}{}

\item{\code{m_AmberAlertCriticalValue}}{}

\item{\code{m_ErrorStringBuffer}}{}

\item{\code{m_ClusteringParametersInitialized}}{}

\item{\code{m_StreamingMode}}{}

\item{\code{m_StreamingModeStatus}}{}

\item{\code{m_ModifiedAt}}{}

\item{\code{m_AnomalyMetricByAnomalyCount}}{}

\item{\code{m_RecentAnomalyCount}}{}

\item{\code{m_ResultsIDArray}}{}

\item{\code{m_TrainingSamples}}{}

\item{\code{m_RecentSamples}}{}

\item{\code{m_RecentRawSamples}}{}

\item{\code{m_RecentTimes}}{}

\item{\code{m_RecentSIs}}{}

\item{\code{m_RecentRIs}}{}

\item{\code{m_RecentADs}}{}

\item{\code{m_RecentAHs}}{}

\item{\code{m_RecentIDs}}{}

\item{\code{m_RecentAMs}}{}

\item{\code{m_RecentAWs}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GetSummaryResponse$new()}}
\item \href{#method-toJSON}{\code{GetSummaryResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{GetSummaryResponse$fromJSON()}}
\item \href{#method-toJSONString}{\code{GetSummaryResponse$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GetSummaryResponse$fromJSONString()}}
\item \href{#method-clone}{\code{GetSummaryResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSummaryResponse$new(
  MagicNumber,
  VersionNumber,
  m_Nano,
  m_BufferStats,
  m_Autotune,
  m_AutotuningElbowClusterCounts,
  m_AutotuningElbowPVArray,
  m_StreamingParameters,
  m_AmberStatus,
  m_Training,
  m_AnomalyThreshold,
  m_AmberWarningCriticalValue,
  m_AmberAlertCriticalValue,
  m_ErrorStringBuffer,
  m_ClusteringParametersInitialized,
  m_StreamingMode,
  m_StreamingModeStatus,
  m_ModifiedAt,
  m_AnomalyMetricByAnomalyCount,
  m_RecentAnomalyCount,
  m_ResultsIDArray,
  m_TrainingSamples,
  m_RecentSamples,
  m_RecentRawSamples,
  m_RecentTimes,
  m_RecentSIs,
  m_RecentRIs,
  m_RecentADs,
  m_RecentAHs,
  m_RecentIDs,
  m_RecentAMs,
  m_RecentAWs
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSummaryResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSummaryResponse$fromJSON(GetSummaryResponseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSummaryResponse$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSummaryResponse$fromJSONString(GetSummaryResponseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetSummaryResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
