% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAmberStatus.r
\name{MAmberStatus}
\alias{MAmberStatus}
\title{MAmberStatus Class}
\description{
MAmberStatus Class

MAmberStatus Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{VersionNumber}}{}

\item{\code{m_Status}}{}

\item{\code{m_Message}}{}

\item{\code{m_Progress}}{}

\item{\code{m_CurrentClusterCount}}{}

\item{\code{m_RetryCount}}{}

\item{\code{m_StreamingWindowSize}}{}

\item{\code{m_TotalInferences}}{}

\item{\code{m_TotalSamplesWritten}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MAmberStatus$new()}}
\item \href{#method-toJSON}{\code{MAmberStatus$toJSON()}}
\item \href{#method-fromJSON}{\code{MAmberStatus$fromJSON()}}
\item \href{#method-toJSONString}{\code{MAmberStatus$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MAmberStatus$fromJSONString()}}
\item \href{#method-clone}{\code{MAmberStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAmberStatus$new(
  VersionNumber,
  m_Status,
  m_Message,
  m_Progress,
  m_CurrentClusterCount,
  m_RetryCount,
  m_StreamingWindowSize,
  m_TotalInferences,
  m_TotalSamplesWritten
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAmberStatus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAmberStatus$fromJSON(MAmberStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAmberStatus$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAmberStatus$fromJSONString(MAmberStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAmberStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
