% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAutotune.r
\name{MAutotune}
\alias{MAutotune}
\title{MAutotune Class}
\description{
MAutotune Class

MAutotune Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{VersionNumber}}{}

\item{\code{m_AutotuningInProgress}}{}

\item{\code{m_PercentComplete}}{}

\item{\code{m_AutotuningSucceeded}}{}

\item{\code{m_NumPatternsToAutotune}}{}

\item{\code{m_ErrorStringBuffer}}{}

\item{\code{m_FeaturesToTuneArray}}{}

\item{\code{m_NCP}}{}

\item{\code{m_AP}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MAutotune$new()}}
\item \href{#method-toJSON}{\code{MAutotune$toJSON()}}
\item \href{#method-fromJSON}{\code{MAutotune$fromJSON()}}
\item \href{#method-toJSONString}{\code{MAutotune$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MAutotune$fromJSONString()}}
\item \href{#method-clone}{\code{MAutotune$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAutotune$new(
  VersionNumber,
  m_AutotuningInProgress,
  m_PercentComplete,
  m_AutotuningSucceeded,
  m_NumPatternsToAutotune,
  m_ErrorStringBuffer,
  m_FeaturesToTuneArray,
  m_NCP,
  m_AP
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAutotune$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAutotune$fromJSON(MAutotuneJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAutotune$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAutotune$fromJSONString(MAutotuneJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAutotune$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
