% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MStreamingParameters.r
\name{MStreamingParameters}
\alias{MStreamingParameters}
\title{MStreamingParameters Class}
\description{
MStreamingParameters Class

MStreamingParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{VersionNumber}}{}

\item{\code{m_EnableAutotuning}}{}

\item{\code{m_SamplesToBufferForAutotuning}}{}

\item{\code{m_GraduationAtRateNumerator}}{}

\item{\code{m_GraduationAtRateDenominator}}{}

\item{\code{m_GraduateAtMaxClusters}}{}

\item{\code{m_GraduateAtMaxSamples}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MStreamingParameters$new()}}
\item \href{#method-toJSON}{\code{MStreamingParameters$toJSON()}}
\item \href{#method-fromJSON}{\code{MStreamingParameters$fromJSON()}}
\item \href{#method-toJSONString}{\code{MStreamingParameters$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MStreamingParameters$fromJSONString()}}
\item \href{#method-clone}{\code{MStreamingParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MStreamingParameters$new(
  VersionNumber,
  m_EnableAutotuning,
  m_SamplesToBufferForAutotuning,
  m_GraduationAtRateNumerator,
  m_GraduationAtRateDenominator,
  m_GraduateAtMaxClusters,
  m_GraduateAtMaxSamples
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MStreamingParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MStreamingParameters$fromJSON(MStreamingParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MStreamingParameters$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MStreamingParameters$fromJSONString(MStreamingParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MStreamingParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
