% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostConfigRequest.r
\name{PostConfigRequest}
\alias{PostConfigRequest}
\title{PostConfigRequest Class}
\description{
PostConfigRequest Class

PostConfigRequest Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{anomalyHistoryWindow}}{}

\item{\code{learningRateNumerator}}{}

\item{\code{learningRateDenominator}}{}

\item{\code{learningMaxClusters}}{}

\item{\code{learningMaxSamples}}{}

\item{\code{featureCount}}{}

\item{\code{streamingWindowSize}}{}

\item{\code{features}}{}

\item{\code{samplesToBuffer}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PostConfigRequest$new()}}
\item \href{#method-toJSON}{\code{PostConfigRequest$toJSON()}}
\item \href{#method-fromJSON}{\code{PostConfigRequest$fromJSON()}}
\item \href{#method-toJSONString}{\code{PostConfigRequest$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PostConfigRequest$fromJSONString()}}
\item \href{#method-clone}{\code{PostConfigRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostConfigRequest$new(
  anomalyHistoryWindow,
  learningRateNumerator,
  learningRateDenominator,
  learningMaxClusters,
  learningMaxSamples,
  featureCount,
  streamingWindowSize,
  features,
  samplesToBuffer
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostConfigRequest$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostConfigRequest$fromJSON(PostConfigRequestJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostConfigRequest$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostConfigRequest$fromJSONString(PostConfigRequestJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostConfigRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
