% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostOutageResponse.r
\name{PostOutageResponse}
\alias{PostOutageResponse}
\title{PostOutageResponse Class}
\description{
PostOutageResponse Class

PostOutageResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{state}}{}

\item{\code{message}}{}

\item{\code{progress}}{}

\item{\code{clusterCount}}{}

\item{\code{retryCount}}{}

\item{\code{streamingWindowSize}}{}

\item{\code{totalInferences}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PostOutageResponse$new()}}
\item \href{#method-toJSON}{\code{PostOutageResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{PostOutageResponse$fromJSON()}}
\item \href{#method-toJSONString}{\code{PostOutageResponse$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PostOutageResponse$fromJSONString()}}
\item \href{#method-clone}{\code{PostOutageResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostOutageResponse$new(
  state,
  message,
  progress,
  clusterCount,
  retryCount,
  streamingWindowSize,
  totalInferences
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostOutageResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostOutageResponse$fromJSON(PostOutageResponseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostOutageResponse$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostOutageResponse$fromJSONString(PostOutageResponseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostOutageResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
