% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostStreamResponse.r
\name{PostStreamResponse}
\alias{PostStreamResponse}
\title{PostStreamResponse Class}
\description{
PostStreamResponse Class

PostStreamResponse Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{state}}{}

\item{\code{message}}{}

\item{\code{progress}}{}

\item{\code{clusterCount}}{}

\item{\code{retryCount}}{}

\item{\code{streamingWindowSize}}{}

\item{\code{totalInferences}}{}

\item{\code{RI}}{}

\item{\code{SI}}{}

\item{\code{AD}}{}

\item{\code{AH}}{}

\item{\code{AM}}{}

\item{\code{AW}}{}

\item{\code{ID}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PostStreamResponse$new()}}
\item \href{#method-toJSON}{\code{PostStreamResponse$toJSON()}}
\item \href{#method-fromJSON}{\code{PostStreamResponse$fromJSON()}}
\item \href{#method-toJSONString}{\code{PostStreamResponse$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PostStreamResponse$fromJSONString()}}
\item \href{#method-clone}{\code{PostStreamResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostStreamResponse$new(
  state,
  message,
  progress,
  clusterCount,
  retryCount,
  streamingWindowSize,
  totalInferences,
  RI,
  SI,
  AD,
  AH,
  AM,
  AW,
  ID
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostStreamResponse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostStreamResponse$fromJSON(PostStreamResponseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostStreamResponse$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostStreamResponse$fromJSONString(PostStreamResponseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PostStreamResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
