% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorUsageInfo.r
\name{SensorUsageInfo}
\alias{SensorUsageInfo}
\title{SensorUsageInfo Class}
\description{
SensorUsageInfo Class

SensorUsageInfo Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{postConfig}}{}

\item{\code{postStream}}{}

\item{\code{putSensor}}{}

\item{\code{getSensor}}{}

\item{\code{getConfig}}{}

\item{\code{getStatus}}{}

\item{\code{getRootCause}}{}

\item{\code{getAmberSummary}}{}

\item{\code{postPretrain}}{}

\item{\code{getPretrain}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SensorUsageInfo$new()}}
\item \href{#method-toJSON}{\code{SensorUsageInfo$toJSON()}}
\item \href{#method-fromJSON}{\code{SensorUsageInfo$fromJSON()}}
\item \href{#method-toJSONString}{\code{SensorUsageInfo$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SensorUsageInfo$fromJSONString()}}
\item \href{#method-clone}{\code{SensorUsageInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorUsageInfo$new(
  postConfig,
  postStream,
  putSensor,
  getSensor,
  getConfig,
  getStatus,
  getRootCause,
  getAmberSummary,
  postPretrain,
  getPretrain
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorUsageInfo$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorUsageInfo$fromJSON(SensorUsageInfoJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorUsageInfo$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorUsageInfo$fromJSONString(SensorUsageInfoJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorUsageInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
